pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[project]
name = "package-name"
version = "0.1.0"
description = "A Python package template"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["python", "template", "package"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0.0",
    "requests>=2.31.0",
    "cachetools>=5.3.0",
    "tenacity>=8.2.0",
    "numpy>=1.21.0",
    "scipy>=1.10.0",
    "networkx>=3.2",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "types-requests>=2.31.0",
]
ml = [
    "lightgbm>=4.0.0",
    "scikit-learn>=1.3.0",
    "shap>=0.42.0",
    "matplotlib>=3.7.0",
]
api = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]
ui = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]

[project.scripts]
package-name = "package_name.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/package-name"
Repository = "https://github.com/yourusername/package-name"
Issues = "https://github.com/yourusername/package-name/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["wikipedia*"]

[tool.setuptools.package-dir]
"" = "."

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
explicit_package_bases = true

# Ignore missing imports for optional dependencies
[[tool.mypy.overrides]]
module = [
    "networkx",
    "pandas",
    "tqdm",
    "lightgbm",
    "matplotlib.*",
    "flask",
    "flask_cors",
    "yaml",
    "colorama",
    "fastapi",
    "fastapi.*",
    "pydantic",
    "uvicorn",
    "tenacity",
    "numpy",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*",
    "venv/*",
    "output/*",
    "examples/*",
]